change long to size_t for frozen pots (other changes for consistency)

+++ cpp_tests/source/test_frozen_hs_wca.hpp
@@ -22,7 +22,7 @@ public:
 	double eps, sca, etrue;
 	Array<double> radii, radii_small, radii_large;
 	Array<double> x, y;
-	Array<long> frozen_dof;
+	Array<size_t> frozen_dof;
 	double* boxvec;
 	virtual void SetUp(){
 		eps=1.0;
